Fix flush cache and Delete cache via caching framework
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/nc_staticfilecache/trunk@88905 735d13b6-9817-0410-8766-e36946ffe9aa

+++ Classes/Cache/StaticFileBackend.php
@@ -91,7 +91,6 @@ class StaticFileBackend extends Typo3DatabaseBackend {
 		} else {
 			parent::set($entryIdentifier, $data, $tags, $lifetime);
 		}
-
 	}
 
 	/**
@@ -175,10 +174,14 @@ class StaticFileBackend extends Typo3DatabaseBackend {
 	 */
 	public function flush() {
 		$absoluteCacheDir = GeneralUtility::getFileAbsFileName($this->cacheDirectory);
-		$tempAbsoluteCacheDir = rtrim($absoluteCacheDir, '/') . '_' . GeneralUtility::milliseconds(TRUE) . '/';
-		rename($absoluteCacheDir, $tempAbsoluteCacheDir);
+		if (is_dir($absoluteCacheDir)) {
+			$tempAbsoluteCacheDir = rtrim($absoluteCacheDir, '/') . '_' . GeneralUtility::milliseconds(TRUE) . '/';
+			rename($absoluteCacheDir, $tempAbsoluteCacheDir);
+		}
 		parent::flush();
-		GeneralUtility::rmdir($tempAbsoluteCacheDir, TRUE);
+		if (isset($tempAbsoluteCacheDir)) {
+			GeneralUtility::rmdir($tempAbsoluteCacheDir, TRUE);
+		}
 	}
 
 	/**
